requirements.txt
apache-airflow==2.8.3
requests
pytest
